Solution: don't use a query candidate
The queryCandidate intermediate variable is not needed. The query class
member can either be a nonempty result or be empty for one of two
reasons:
- Empty from initial setup, no filter was ever applied
- Empty result of an applied filter

Both empty results are a Query object in the same state (i.e. empty
vector of filters (Query::filters) and empty package set
(Query::result)), so the query can be modified in-place.

Closes: #795
Approved by: <try>